much better organization and logging
