+++ app.js
@@ -2,7 +2,7 @@
 require("dotenv").config({ path: "../config/curbmap.env" });
 const express = require("express");
 const path = require("path");
-const logger = require("morgan");
+const morgan = require("morgan");
 const passport = require("passport");
 const LocalStrategy = require("passport-local").Strategy;
 const cookieParser = require("cookie-parser");
@@ -33,7 +33,6 @@ var whitelist = [
 ];
 var corsOptions = {
   origin: function(origin, callback) {
-    winston.log("info", "origin: ", origin);
     if (whitelist.indexOf("*") !== -1 || whitelist.indexOf(origin) !== -1) {
       callback(null, true);
     } else {
@@ -45,7 +44,23 @@ var corsOptions = {
 app.options("*", cors(corsOptions)); // include before other routes
 app.use(cors(corsOptions));
 app.use(compression());
-app.use(logger("dev"));
+morgan.token('id', function getId(req) {
+    return req.id
+});
+const loggerFormat = ':id [:date[web]] ":method :url" :status :response-time ms';
+app.use(morgan(loggerFormat, {
+    skip: function (req, res) {
+        return res.statusCode < 400
+    },
+    stream: process.stderr
+}));
+
+app.use(morgan(loggerFormat, {
+    skip: function (req, res) {
+        return res.statusCode >= 400
+    },
+    stream: process.stdout
+}));
 app.use(cookieParser());
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(bodyParser.json());
@@ -53,57 +68,57 @@ app.use(bodyParser.json());
 app.use(express.static(path.join(__dirname, "public")));
 
 function shouldCompress(req, res) {
-  if (req.headers["x-no-compression"]) {
-    // don't compress responses with this request header
-    return false;
-  }
-
-  // fallback to standard filter function
-  return compression.filter(req, res);
+    if (req.headers["x-no-compression"]) {
+        // don't compress responses with this request header
+        return false;
+    }
+    
+    // fallback to standard filter function
+    return compression.filter(req, res);
 }
 
 app.use(passport.initialize());
 
 passport.serializeUser(function(user, cb) {
-  cb(null, user.username);
+    cb(null, user.username);
 });
 
 passport.deserializeUser(function(username, cb) {
-  findUser(username, cb);
+    findUser(username, cb);
 });
 
 // We will add other Strategies, such as FB strategy
 passport.use(
-  new JWTStrategy(
-    {
-      jwtFromRequest: ExtractJWT.fromAuthHeaderAsBearerToken(),
-      secretOrKey: TOKEN_KEY,
-      algorithms: ["RS384"]
-    },
-    (jwtPayload, cb) => {
-      return cb(null, jwtPayload);
-    }
-  )
+    new JWTStrategy(
+        {
+            jwtFromRequest: ExtractJWT.fromAuthHeaderAsBearerToken(),
+            secretOrKey: TOKEN_KEY,
+            algorithms: ["RS384"]
+        },
+        (jwtPayload, cb) => {
+            return cb(null, jwtPayload);
+        }
+    )
 );
-const main = require("./routes/main")
-app.use("/", main);
+const v1 = require("./routes/v1");
+app.use("/v1", v1);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
-  var err = new Error("Not Found");
-  err.status = 404;
-  next(err);
+    var err = new Error("Not Found");
+    err.status = 404;
+    next(err);
 });
 
 // error handler
 app.use(function(err, req, res, next) {
-  // set locals, only providing error in development
-  res.locals.message = err.message;
-  res.locals.error = req.app.get("env") === "development" ? err : {};
-
-  // render the error page
-  res.status(err.status || 500);
-  res.render("error");
+    // set locals, only providing error in development
+    res.locals.message = err.message;
+    res.locals.error = req.app.get("env") === "development" ? err : {};
+    
+    // render the error page
+    res.status(err.status || 500);
+    res.render("error");
 });
 
 module.exports = app;
