+++ package.json
@@ -55,8 +55,11 @@
     "serve-favicon": "^2.5.0",
     "sharp": "^0.20.2",
     "supertest": "^3.0.0",
+    "swagger-jsdoc": "^1.10.3",
+    "swagger-ui-express": "^3.0.10",
     "twilio": "^3.16.0",
     "uuid": "^3.2.1",
-    "winston": "^2.4.2"
+    "winston": "^2.4.2",
+    "yamljs": "^0.3.0"
   }
 }
