--- routes/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  init: require("./main.js")
-};
